Fixes #43.
Cause: a list of atomic length scales is used when creating an
adjacency rule. The list was cache with the functions.lru_cache, which
returns a reference to the same mutable array instead of a new copy.
As a result, this list gets modified by the `zoom` argument, leading to
inconsistent results when `graphdot.experimental.metric.M3` is
instantiated repeatedly.